requirements.txt
python-docx
PyPDF2
scikit-learn
tqdm
streamlit